fix: move SSRF module to API package to fix Next.js build
The shared package is imported by the web app, and node:dns/promises
cannot be bundled by webpack. Move ssrf.ts to apps/api/src/utils/
since only server-side code needs it.
