+++ packages/shared/src/index.ts
@@ -17,7 +17,6 @@ export * from "./error-classifier.js";
 export * from "./types/session.js";
 export * from "./types/mcp.js";
 export * from "./utils/off-peak.js";
-export * from "./utils/ssrf.js";
 export * from "./types/optio-settings.js";
 export * from "./types/optio-action.js";
 export * from "./optio-tools.js";
